# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: QUTCraft CMS API
  version: 0.1.0
  description: |
    公开门户与受认证后台共享的 API 契约。门户只读取已发布、可公开的数据；草稿、成员隐私与审核能力仅通过独立的管理前缀暴露。
    本文件是 Swagger UI 与 Apifox 的共同导入源；公开接口无认证，管理接口接受 Bearer JWT 或第一方网页的 HttpOnly Access Cookie，并继续执行 RBAC 授权。
  license:
    name: Proprietary
    url: https://github.com/QUTCraft/qutc-platform
servers:
  - url: https://api.qutcraft.local
    description: 示例部署地址；由 Apifox 环境或运行环境变量覆盖
security: []
...
paths:
...
  /api/v1/invitations/{token}:
    get:
      tags: [Auth]
      operationId: getInvitation
      summary: 获取邀请链接公开信息
      description: 只返回组织名称、邀请邮箱、角色和有效期；token 只作为路径中的一次性凭据，不会在响应中回显。
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, minLength: 32, maxLength: 128 }
      responses:
        '200':
          description: 邀请信息
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: 邀请链接已过期、撤销或已使用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/invitations/{token}/accept:
    post:
      tags: [Auth]
      operationId: acceptInvitation
      summary: 当前登录用户接受邀请
      description: 当前 Bearer JWT 的邮箱必须与邀请邮箱一致；接受时在事务中激活成员关系、授予邀请角色并将 token 标记为已使用。
      security: [{ BearerAuth: [] }, { AccessCookie: [] }]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, minLength: 32, maxLength: 128 }
      responses:
        '200':
          description: 邀请已接受
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationAcceptanceResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: 邀请链接已过期、撤销或已使用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/v1/admin/invitations:
    get:
      tags: [Admin]
      operationId: listAdminInvitations
      summary: 获取当前组织的成员邀请
      description: 仅返回邀请元数据和邮件投递状态，不返回或恢复任何明文邀请 token；可按邀请状态筛选。
      security: [{ BearerAuth: [] }, { AccessCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/InvitationStatus' }
      responses:
        '200':
          description: 当前组织的邀请列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminInvitationListResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      operationId: createAdminInvitation
      summary: 创建成员邀请
      description: 同一组织同一邮箱只能有一个待处理邀请；创建邀请时会同时预创建 invited 状态的账户和成员关系，账户在设置密码并接受邀请前不能登录。邀请只允许授予 member、editor 或 administrator，不能授予 owner。
      security: [{ BearerAuth: [] }, { AccessCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvitationRequest' }
      responses:
        '201':
          description: 邀请及待激活账户已创建；invite_url 中包含仅返回一次的明文 token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateInvitationResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/admin/invitations/{invitation_id}:
    delete:
      tags: [Admin]
      operationId: revokeAdminInvitation
      summary: 撤销待处理的成员邀请
      description: 仅可撤销当前组织中仍为 pending 的邀请。撤销后原 token 立即失效，邀请记录与审计记录保留。
      security: [{ BearerAuth: [] }, { AccessCookie: [] }]
      parameters:
        - name: invitation_id
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200':
          description: 邀请已撤销
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 后台访问令牌；服务端必须基于组织成员身份与 RBAC 进行二次授权。
    AccessCookie:
      type: apiKey
      in: cookie
      name: qutc_access
      description: 第一方网页使用的短期 HttpOnly Access Cookie；与 Bearer JWT 使用相同的签名、组织隔离和 RBAC 校验。
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    NotFound:
      description: 当前组织不存在或没有公开门户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 缺少、失效、无法验证的访问令牌，或当前账户/组织成员关系已停用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 当前角色没有执行此管理操作的权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: 请求数据未通过校验
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源状态已发生变化，无法重复执行操作
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    TooManyRequests:
      description: 当前客户端对该接口的请求超过限额
      headers:
        Retry-After:
          description: 建议等待的秒数
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Limit:
          description: 当前窗口允许的最大请求数
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Remaining:
          description: 当前窗口剩余请求数
          schema: { type: integer, minimum: 0 }
        X-RateLimit-Reset:
          description: 当前窗口结束的 Unix 时间戳
          schema: { type: integer, format: int64 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    Meta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string, example: req_01J0QUTC9B7ABCD }
    ListMeta:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          required: [page, page_size, total]
          properties:
            page: { type: integer, minimum: 1, example: 1 }
            page_size: { type: integer, minimum: 1, maximum: 100, example: 20 }
            total: { type: integer, minimum: 0, example: 3 }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: portal.organization_not_found }
            message: { type: string, example: 未找到可公开展示的组织。 }
            details:
              oneOf:
                - { type: object, additionalProperties: true }
                - { type: array, items: { type: object, additionalProperties: true } }
            request_id: { type: string, example: req_01J0QUTC9B7ABCD }
...
    InvitationRole:
      type: string
      enum: [member, editor, administrator]
    InvitationStatus:
      type: string
      enum: [pending, accepted, expired, revoked]
    Invitation:
      type: object
      required: [id, organization_id, organization_name, email, role, status, expires_at, created_at]
      properties:
        id: { type: string }
        organization_id: { type: string }
        organization_name: { type: string, maxLength: 160 }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/InvitationRole' }
        status: { $ref: '#/components/schemas/InvitationStatus' }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    CreateInvitationRequest:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        role: { $ref: '#/components/schemas/InvitationRole' }
        expires_in_hours: { type: integer, minimum: 1, maximum: 720, default: 168 }
    CreateInvitationResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/Invitation'
            - type: object
              required: [invite_url, delivery]
              properties:
                invite_url: { type: string, format: uri-reference }
                delivery: { $ref: '#/components/schemas/EmailDelivery' }
        meta: { $ref: '#/components/schemas/Meta' }
    EmailDeliveryStatus:
      type: string
      enum: [disabled, pending, sent, failed]
    EmailDelivery:
      type: object
      required: [status, adapter, attempts]
      properties:
        status: { $ref: '#/components/schemas/EmailDeliveryStatus' }
        adapter: { type: string, enum: [disabled, smtp] }
        attempts: { type: integer, minimum: 0 }
        last_error: { type: string, maxLength: 500 }
        last_attempt_at: { type: string, format: date-time }
        sent_at: { type: string, format: date-time }
    AdminInvitation:
      allOf:
        - $ref: '#/components/schemas/Invitation'
        - type: object
          required: [delivery]
          properties:
            delivery: { $ref: '#/components/schemas/EmailDelivery' }
    AdminInvitationListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AdminInvitation' }
        meta: { $ref: '#/components/schemas/ListMeta' }
...
    InvitationResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Invitation' }
        meta: { $ref: '#/components/schemas/Meta' }
    InvitationAcceptanceResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/Invitation'
            - type: object
              required: [membership_id]
              properties:
                membership_id: { type: string }
        meta: { $ref: '#/components/schemas/Meta' }
...